Add property support for currently implemented Label features
- Introduce LabelPropertyIndex entries for supported APIs
- Implement PropertyHandler for SetProperty/GetProperty routing

+++ dali-ui-foundation/public-api/label.autogen.h
@@ -145,7 +145,7 @@
   * \
   * @param[in] gap The gap in pixels. \
   */ \
-  ChildClass& SetMarqueeGap(float gap) { Label::SetMarqueeGap(gap); return *this; } \
+  ChildClass& SetMarqueeGap(int gap) { Label::SetMarqueeGap(gap); return *this; } \
   /** \
   * @brief Sets how the marquee stops. \
   * \
